Major iteration across Rig for multi-process architecture
These changes affect a lot of Rig and although ideally, I would have
preferred to split this work into smaller patches, most of the changes
are inter-related and so it would take quite a lot of effort to split up
now.

Note: this work is still not complete, since for example, not all of the
edit operations in rig-engine-op.c have been implemented yet. The idea
is to first verify that the property-set operation works before
completing the other operations.