Move RutEntity under rig/ as RigEntity
This is a pretty huge re-arrangement because of the various dependencies
on RutEntity. All of the components have moved as well as RutAsset,
RutAssetInspector, RutPropInspector, RutInspector, RutDepthOfField
and RutDownsampler.

This was done because we need more specialized garbage collection for ui
objects, tightly coupled with rig. More generally it has also seemed a
bit inconvenient to develop entities, components and assets as rut/ apis
when they they are so central to a rig ui.