Add "reap" mechanism, marking obj's for garbage collection
This introduces a mechanism for queuing an object to be garbage
collected the next time rig_engine_garbage_collect() is called as well
as any objects owned by the given object.

This mechanism is use by simulator__load to ensure that all the objects
of an old ui can be garbage collected and unregistered before loading a
new ui.

This mechanism will also be required for any engine OP that is meant to
destroy something, so when UI logic deletes something then it will be
properly garbage collected.

This also updates the way that we apply operations in rig-engine-op.c to
replace apply_ctx->delete_object_cb with an unregister_id_cb instead.
The apply callbacks themselves will no be responsible for reaping
objects.  The unregster_id_cb is used by the frontend for potentially
unregistering temporary ids.