shell: Grab pointer on X11 when grabbing input
This adds a new rut_shell_grab_pointer() api that uses the extra
semantic information to start an explicit pointer grab if running on X11
so that we can continue to receive events when the mouse moves outside
our window, until all buttons have been released. The grab is implicitly
released when all buttons have been released so there's currently no api
for explicitly canceling a pointer grab.

Currently this new api is only used in rut-number-slider.c

Note: That this requires a patch to SDL so you need to build against the
rig branch of SDL at https://github.com/rig-project/sdl