rpc: free message on closure free
This updates the rpc layer so that the unpacked message that gets passed
to us when servicing a request doesn't get freed immediately on return
from the service function, and instead gets freed when that requests
corresponding closure function is called, signalling that the request
has been handled.

This is necessary for rig-slave to be able to batch edit messages until
it is ready to send them to the simulator.