slave: fix loading and editing the ui
This ensures that the ids we use for loading a ui in a slave device
correspond to edit-mode object pointers within an editor. Instead of
applying edit operations in a slave, separate from where we forward the
same operations to the simulator, we now apply slave operations in sync
with the rate that they are sent to the slave's simulator.