updates linux and osx build scripts and starts android build script
This updates the scripts for building linux and osx binaries of the Rig
editor.

This also adds an initial script for building a set of
PREBUILT_SHARED_LIBRARY modules for Android that can be used for
running Rig on an android device.

This also adds an android/project that contains an Android project that
can be used to build a rig-slave package.