xlib: remove deprecated apis
This removes some deprecated cogl_xlib_ code that depends on a global
CoglContext. Note: we keep cogl-xlib.h as a top-level header for xlib
specific apis. Since we now have no longer include cogl-xlib.h
internally this slightly simplifies the guarding inside the header to
trigger an error from now on if the header is use internally, as we do
for cogl.h.