Remove support for GL's legacy fixed function apis
This removes the GL fixed function vertend, fragend and progend since we
don't have any need to support non programmable GPUs and since it
reduces the number of code paths we maintain.

This adds "nop" backends so that for now, we continue to keep support
internally for selecting pipeline backends at runtime.