frontend: enable binding to abstract socket
This enables the frontend to bind to an abstract unix socket that
listens for a connection from a simulator instead of forking and
exec()ing the simulator. This mechanism is always used on Android but it
can also optionally be used on linux since it is very convenient for
debugging.

To choose to bind to an abstract socket _RIG_USE_ABSTRACT_SOCKET should
be set in the environment. The same variable should be set when running
rig-simulator to make it connect to that socket.