consolidate mainloop code
This adds a rut_poll_ api that provides a wrapper for any mainloop
functionality that we need within Rut and Rig. In particular it tracks
what file descriptors we want to poll, lets code register idle functions
and handles integration with Cogl's polling api.

This removes the mainloop implementation in rig-protobuf-c-dispatch.c
and instead layers that code on this rut_poll_ api.

This adds a glib mainloop source that essentially acts as the backend
for this rut_poll_ api.