fragend-glsl: avoid re-getting the shader state
This reduces how often we have to get the shader state for a pipeline
while generating the code for each layer of the pipeline. This also
separates looking up the corresponding layer state with a layer index
from the codegen code. This would allow us to avoid redundant index
based lookups if we were to start generating code for layers that aren't
used internally.