undo-redo: disable test undo-redo for every op
To help us catch bugs with undo-redo, after applying any operation we
would then immediately undo the operation, and redo it, so if undo and
redo weren't working then it wouldn't be possible to do an operation in
the first place.

The problem is that the deletion of objects is now handled lazily such
that we only unregister object ids of deleted objects when we return to
the mainloop and we will hit debug assertions if we try and re-register
an object trying to give it multiple identifiers.