re-locate more engine code
This moves more code out of rig-engine.c into the frontend, simulator,
device, slave and editor.

A problem with RigEngine is that has become a dumping ground for state
that needs to be accessible across most Rig components. The aim now is
to take advantage of the rut trait system and create separate
program-engine types, like RigEditorEngine and RigDeviceEngine that
can be made up of traits and it will be easier for to re-jig the
organisation of code that supports the engines, without having just one
monolythic type.