rpc: split out code for opening listening fd
This splits out the code for opening a listening fd for an rpc server
into rut-os.c. This lets us construct a server and give it an already
listening fd that can be created separately and lets us share the api
for opening a listening abstract socket that may also be used for
setting up peer to peer ipc between a frontend and simulator.