Fix problem with mapping and forwarding ops
Previously we weren't mapping the IDs of new objects, which made sense
considering we want to use those IDs to register the object and until
the object is registered then there's no way to map the ID.

The problem though is that in the editor we map operations from the edit
mode ui to the play mode ui and then forward edits to the simulator.

When mapping to the play mode ui the IDs correspond to edit mode objects
so when new play objects are registered we are able to associate them
with the corresponding edit mode object. When we were then forwarding to
the simulator though we still had edit mode IDs for those mapped play
mode operations which was resulting in duplicate IDs being sent to the
simulator.

This updates how operation mapping is handled so that we apply
operations at the same time which gives the mapping code an opportunity
to finish mapping an operation after it has been applied (and any new
objects have been registered with the current ID)