Allow running frontend + simulator in same thread
This adds support for running the simulator in the same thread as the
frontend, sharing the same mainloop.

The mainloop is shared by introducing the idea of a "main" shell and a
simulator can optionally be constructed in association with a main shell
so that when it creates its own shell and calls into rut_poll_init()
then rut-poll.c will know not to create a separate mainloop and
when rut_poll_run() is called by rig_simulator_run() it also won't
actually block and run the mainloop; only calling rut_poll_run() with
the main shell will block.

Communication between the frontend and the simulator will still happen
via rpc-peer end points but instead of writing to a file descriptor,
data will be written to a rig-pb-c-data-buffer.

This mode of operation will initially be used on Android, and is also
expected to be used with Emscripten too which doesn't support threads.