Fold rut_context_t object into rut_shell_t
The separation of state between context and shell objects is currently
arbitrary. In particular it's awkward having the Cogl context be in the
rut-context since the setup of the context depends on the platform,
which is largely what the shell helps abstract. This is a problem a.t.m
because adding support for Android needs much more careful control over
the initialisation of the shell/context state which is difficult to
manage with the state split up.

It also generally simplifies things a bit to fold this state into one
object.

In making this change this removes the rut.c dumping ground which
included some context apis that have moved to rut-shell.c; some settings
apis, now split into rut-settings.c and some texture-cache code, now
split into rut-texture-cache.c.