rpc-network: Fix incorrect copyright headers
There was a leftover copyright header relating to the protobuf-c project
since rig-rpc-network.c used to have a copy of the dispatching code from
protobuf-c that was removed in 55ca089892d6d204ac82ece07d3973ec02072b4b

There was also a leftover LGPL header since Rig was re-licensed. This
code has been stripped down since it was first introduced to help
integrate the protobuf-c-rpc api and the aim is to remove the need for
this entirely at some point. What's currently left is just a trivial
shim over the protobuf-c-rpc api written by me (copyright Intel
Corporation) so it looks fine to update the header now.