rut-poll: Fix copyright header
It was an oversight that the rut-poll.h header was added with an LGPL
copyright notice. The code landed soon after the codebase was
re-licensed but I expect it was started on a branch before the
re-license. It was also originally based on code (copyright Intel
Corporation) from Cogl that has also been re-licensed under the MIT
license.