carray: use growth exponent of 1.5
This switches carray to using a growth exponent of 1.5 instead 2 so
in situations where an array is repeatedly grown there is a better
chance of being able to incorporate the memory of earlier sizes.

I can't say I've scrutinised this myself but based on this write up for
example it seems likely to be better:

https://github.com/facebook/folly/blob/master/folly/docs/FBVector.md

Obviously tweaking an exponent is a trivial thing to do and fingers
crossed the new number is better for something, but it could also make
sense for us to depend on a specific allocator (probably jemalloc) and
look at real metrics.