clib: remove filenames from source + 'eglib'
The clib code started as a copy of eglib from the Mono project with
files starting with 'g' instead of 'c'. In general it doesn't seem
useful to duplicate the filename in the source and since the names have
changed now this patch strips the declared filenames. Similarly this
removes some out of date references to the 'eglib' project name.