camera-view: simplify camera management
Instead of connecting the viewing camera off of multiple other
entities that were meant to help with the user navigation we
now just make the camera a direct child of the scene which
can be freely positioned and oriented.

Another notable change is that rig-renderer.c has more complete control
over the separate rendering passes involved in rendering which should
help if we look at creating alternative renderers later.

XXX: this breaks existing .rig content because there isn't currently a
way to flip everything vertically. (trying to do it by moving the camera
would either also flip the x-axis or the z-axis and complicate the
default positioning of the camera.)