slave: intercept F11 to fullscreen
FIXME: this doesn't work because slave input events are only dispatched
in the simulator.

It seems like it would be hacky to just intercept F11 in the frontend so
instead we should look at fleshing out the simulator runtime support,
including adding some protocol + api for requesting the frontend to
fullscreen a window.