cogl: remove journal and primitives code
Deletes *lots* of code for batching rectangles that's not as useful as
it used to be and given the way it's integrated throughout Cogl, by
removing it, this should make it easier to refactor and simplify other
aspects of cogl later too.

For batching rectangles it's now recommended that it should be dealt
with at a higher level, most likely based on instanced attributes.