Start adding support for emscripten/html5
This makes lots of emscripten portability updates and support
throughout clib, cogl, rut and rig

For building the necessary dependencies such as freetype with the
emscripten toolchain this adds a build/emscripten/build.sh script.
Prebuilt bitcode can then be distributed along with Rig for compiling
and packaging a html5 based application at runtime.

So instead of adding emscripten to our autoconf/automake build scripts
it's more important that native builds of Rig can be distributed with
source and an application built with Rig can be compiled and packaged
for deployment as a html5 application.

For now this starts by adding vanilla gnu Makefiles that will assume
pre-requisites such as freetype or harfbuzz have already been built via
build/emscripten/build.sh and just support compiling clib, cogl, rut and
rig with appropriate, application specific compilation options.

For testing, this adds a minimal shell-hello test for testing
rut-shell without rig

This adds an initial rig-package-html5 helper script

This script can be distributed with Rig so that developers can compile
and package an application for distribution as a html5 application. This
is simply a prototype script for now. Eventually it should be possible
to control packaging from the Rig ui and any text based tool should
be updated to be more portable.