cogl/fragend-glsl: be more careful iterating layers list
While generating the layers, each current layer is removed from the list
of layers before recursively generating dependencies.

We were using the c_list_for_each_safe macro to iterate the list but
that can only handle the current link being removed from the list while
in this case it's possible that by the time we try and iterate our
loop again the list is empty and the temp pointer for the "next" link is
now invalid.

Loop iteraction no longer reads ahead for a pointer that may become
invalid and now terminates when the list is empty.