rig: start adding native-module component + editor2
The aim is to be able to reimplement the editor as a rig based ui and
as a step towards that this adds a component that can load a DSO so
native code can e.g. hook into input handling and frame updates with
the idea that some of the existing editor code can be de-coupled from
Rig internals and instead loaded via this native module component.