lots of input handling updates
In general scene picking has moved from rig-camera-view to rig-ui
considering that we need to handle input in the simulators that
no longer have a rig-camera-view.

A lot of input handling complexity has also been deleted now that
we only plan to support input for entities as we stop using the
the 2d ui infrastructure in rut.