stash: replace r_open_view with _new/_delete apis
We need an object for each view in the simulator and frontend and to
instead of r_open_view being an action that only creates an object in
the frontend r_view_new() constructs an object in the simulator that has
"width" + "height" properties that are kept in sync with the
corresponding properties in the frontend view object. Like entities and
components an explicit _delete api is given without exposing a ref
counting api.

Notably the width + height properties are needed to be able to update
the projection matrix of cameras in the simulator with the correct
aspect ratio.